refactor(e2e): extract interaction helpers into test/e2e/helpers modules
Move the reusable ExTester interaction helpers out of the single large test file
into focused, reusable modules under test/e2e/helpers/:

- constants            timeouts + the output-iframe selector
- fixtures             copyFixtureToTempDir
- notifications        dismissAllNotifications, waitForNotification
- quickInput           tryOpenInputBox, clickDialogOkButton
- workspace            openFolderViaDialog, openWorkspaceFile
- notebook             clickRunAll, readRenderedOutput, runAndAwaitOutput
- deepnoteEnvironment  createEnvironment, selectEnvironmentForNotebook
- index                barrel re-export

Helpers take the WebDriver from VSBrowser.instance.driver and receive the
notebook name as a parameter instead of closing over suite state, so future
suites can reuse them. helloWorld.e2e.test.ts shrinks from 439 to 87 lines and is
now just the suite wiring. Behaviour is unchanged — verified passing.
