+++ test/e2e/helpers/index.ts
@@ -1,0 +1,8 @@
+// Barrel re-export so suites can `import { … } from '../helpers'`.
+export * from './constants';
+export * from './deepnoteEnvironment';
+export * from './fixtures';
+export * from './notebook';
+export * from './notifications';
+export * from './quickInput';
+export * from './workspace';
